+++ src/lib/sidebar.test.ts
@@ -47,7 +47,7 @@ describe("Sidebar keyboard activation", () => {
       onNavigate: (page: number) => navigated.push(page),
       results: { onPick: () => {} },
       comments: { onPick: () => {} },
-      marks: { onPick: () => {} },
+      marks: { onPick: () => {}, onRemove: () => {} },
       pages: {
         doc: 1,
         pageCount: 40,
@@ -98,7 +98,7 @@ describe("Sidebar keyboard activation", () => {
       onNavigate: () => {},
       results: { onPick: () => {} },
       comments: { onPick: () => {} },
-      marks: { onPick: () => {} },
+      marks: { onPick: () => {}, onRemove: () => {} },
       pages: {
         doc: 1,
         pageCount: 40,